% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dicho.R
\docType{data}
\name{dichotomous_response}
\alias{dichotomous_response}
\title{dichotomous item responses}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(dichotomous_response)
}
\description{
A simulated dataset for dichotomous item responses. Percent correct of responses ranges from 0.2 to 0.95.
N = 100. Number of items = 10.
}
\examples{

data(dichotomous_response)
head(dichotomous_response)

}
\references{
This data set was artificially created for the CTTvis package.
}
\keyword{datasets}
