% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-reliability.R
\docType{data}
\name{reliability_df}
\alias{reliability_df}
\title{reliability dataframe}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(reliability_df)
}
\description{
A simulated dataset for reliability analysis testing.
Each item varies in terms of their influence on the overall reliability if dropped from the test.
N = 100. Number of items = 10.
}
\examples{

data(reliability_df)
head(reliability_df)

}
\references{
This data set was artificially created for the CTTvis package.
}
\keyword{datasets}
